pyproject.toml
[tool.poetry]
name = "easymenu"
version = "0.1.0"
description = "A small Python/Django web application to easily create a food or drink menu in multiple languages."
authors = ["jiri.one"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.3"
psycopg2-binary = "^2.9.5"
django-money = "^3.0.0"
beautifulsoup4 = "^4.11.1"
django-modeltrans = "^0.7.3"


[tool.poetry.group.tests.dependencies]
pytest = "^7.2.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
